1. 移除不必要的版权声明和包含的头文件，简化代码结构。 2. 在检测预处理线程中添加处理时间的日志记录，优化性能监控。 3. 更新推流线程的分辨率设置，以提高性能。 4. 在推流过程中添加处理时间的日志记录，便于性能分析。 5. 使用更快的图像缩放算法（INTER_NEAREST）以提升处理速度。

+++ src/main.cpp
@@ -20,7 +20,6 @@
 #include "AclLiteApp.h"
 #include "AclLiteResource.h"
 #include "AclLiteThread.h"
-#include "AclLiteType.h"
 #include "AclLiteUtils.h"
 #include "Params.h"
 #include "dataInput/dataInput.h"
@@ -34,7 +33,6 @@
 #include <fstream>
 #include <iostream>
 #include <json/json.h>
-#include <map>
 
 using namespace std;
 namespace
@@ -298,7 +296,7 @@ void StartApp(AclLiteResource &aclDev)
 int main(int argc, char *argv[])
 {
     // Check input args: the path of configure json
-    if ((argc != argNum) || (argv[1] == nullptr))
+    if ((argc != static_cast<int>(argNum)) || (argv[1] == nullptr))
     {
         ACLLITE_LOG_ERROR("Please input: ./main <json_dir>");
         return ACLLITE_ERROR;
